transformers/appairflow6/requirements.txt
apache-airflow==2.2.3
azure-storage-file-share==12.5.0
azure-storage-blob==12.8.1
pandas==1.3.3
pyyaml==5.4.1
